feat: Add new ValorSensor construct
Allow sensors to be their own object, which simplifies
the logic in the subsystems.

This also allows us to re-use common sensor types year over year.
For example the current sensing circular buffer on our intake
is a good sensor to use in the future, therefore create
a ValorCurrentSensor that can be used by multiple subsystems