feat: ValorControllers
ValorController is an abstract class that has 2 implementations:
* ValorNeoController
* ValorFalconController

Each controller houses the logic required to drive either a
falcon or a neo. This is useful as we forgot to call specific
functions in different subsystems that are associated with those motors.

Instead, all of the logic is the same for each motor but can be
modified if need be. There are getters and setters for the important
functions for the motors.

Another thing to note is that since ValorController is an abstract class,
the implementations share the exact same function calls. Using either class,
you won't have to worry about which motor is under the hood since the calls
all look the same.

Most importantly, this synchronizes the units of the motors. The falcon
and neo both use different units at the base level. This ValorController
class now makes sure that to Valor students, both have the same units